+++ changelog.md
@@ -1,14 +1,29 @@
-version 1.9
+
+------------------------------------------------------
+Mod version 2.0 for game version 0.32.1 | 2025-04-28
 ------------------------------------------------------
 
-Added 4 new pets to the game:
+Increase the swim speed buff from 20% to 45%
+
+Added 3 new craftable companions:
+
+- Duck (swim speed)
+- Fox Cub (armor penetration)
+- Snow Hare (max resilience)
+
+------------------------------------------------------
+Mod version 1.9 for game version 0.32.1 | 2025-04-28
+------------------------------------------------------
+
+Added 4 new craftable companions:
 
 - Honey Bee (attack speed)
 - Phoenix (critical damage)
 - Stabby Bush (critical chance)
 - Turtle (max mana)
 
-version 1.8
+------------------------------------------------------
+Mod version 1.8 for game version 0.32.1 | 2025-04-27
 ------------------------------------------------------
 
 First release of the mod